refactor: organize file structure for DetailPage
