+++ client/components/page/Detail/index.ts
@@ -1,0 +1,1 @@
+export { default as DetailPage } from './Detail.page';
